requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
pyyaml==6.0.2
litellm==1.48.7
pydantic==2.9.2
pytest==8.3.3
sqlalchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
aiohttp==3.10.5
black==24.8.0
psycopg2-binary==2.9.9